requirements.txt
# To ensure there exists a wheel for win32/py27
cython<0.29.8;python_version<="2.7"
# No requirements going forward
cython;python_version>="3"
hidapi
pyserial
